Stylesheet needs to be compatible with cached HTML
The mw-portlet class is new. When we originally worked on these
changes we assumed a week cadence between branches however due
to delays in the train last week what actually happened was the train
rolled out a few days after the previous branch.

Given the change in train the CSS must be compatible with cached HTML.
This can be reverted in a weeks time.

Bug: T265543
Change-Id: I4ab36800e103d518fdef6af0ed5b10136fa788e5